refactor: a bit
