+++ client/packages/fast-writer/.env-example.t0rn
@@ -7,4 +7,4 @@ export RANGE_INTERVAL_SECS=12
 
 export LOG_LEVEL=debug
 export LOG_PRETTY=true
-export SIDEEFFECTS='[{"target":"roco","asset":null,"maxReward":1,"insurance":1,"action":"transfer","targetAccount":"5F3vTGGG56DetArfvmtZseV6fEZeCQpdYgijzzsQdBqdxmen","amount":1,"enforceExecutor":null,"rewardAsset":null,"count":1000,"txType":"single"}]'
+export SIDEEFFECTS='[{"target":"roco","asset":null,"maxReward":1,"insurance":1,"action":"transfer","targetAccount":"5F3vTGGG56DetArfvmtZseV6fEZeCQpdYgijzzsQdBqdxmen","amount":1,"enforceExecutor":null,"rewardAsset":null,"count":10,"txType":"single"}]'
