+++ client/packages/fast-writer/src/services/fast-writer/enums.ts
@@ -1,0 +1,4 @@
+export enum TxType {
+  Single = 'single',
+  Batch = 'batch',
+}
